backend/services/chat-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
openai>=1.30.0
sse-starlette==1.6.5
httpx>=0.25.0

